% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_col.R
\name{boxplot_col}
\alias{boxplot_col}
\title{Obtiene diagramas de caja de cada columna}
\usage{
boxplot_col(x, y)
}
\arguments{
\item{x}{A vector o dataset}

\item{y}{A logical if is TRUE elimina outliers}
}
\value{
histogramas
}
\description{
Obtiene diagramas de caja de cada columna
}
\examples{
boxplot_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)),F)
}
