% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conteoNA.R
\name{conteo_NAS}
\alias{conteo_NAS}
\title{Cuenta los NAS}
\usage{
conteo_NAS(x)
}
\arguments{
\item{x}{A vector o dataset}
}
\value{
Total NAS por columna
}
\description{
Cuenta los NAS
}
\examples{
conteo_NAS(data.frame(x=c(1, 3, NA, NA, 6),y=c("A", NA, "C", "D", "E"),z=c(NA, TRUE, NA, TRUE, NA)))
}
