% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_col.R
\name{cv_col}
\alias{cv_col}
\title{Obtiene el coeficiente de variacion por columnas
Convertir a data frame si se obtiene variables categoricas falsamente}
\usage{
cv_col(x)
}
\arguments{
\item{x}{A vector o dataset}
}
\value{
vector coef de variacion por columna
}
\description{
Obtiene el coeficiente de variacion por columnas
Convertir a data frame si se obtiene variables categoricas falsamente
}
\examples{
cv_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)))
}
