% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delcategoric_col.R
\name{delcategoric_col}
\alias{delcategoric_col}
\title{elimina columnas con datos no numericos
install.packages("dplyr")
library(dplyr)}
\usage{
delcategoric_col(x)
}
\arguments{
\item{x}{A vector o dataset}
}
\value{
data con solo numeric
}
\description{
elimina columnas con datos no numericos
install.packages("dplyr")
library(dplyr)
}
\examples{
delcategoric_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)))
}
