% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_col.R
\name{hist_col}
\alias{hist_col}
\title{Obtiene los histograma de cada columna
Convertir a data frame si se obtiene variables categoricas falsamente}
\usage{
hist_col(x, y)
}
\arguments{
\item{x}{A vector o dataset}

\item{y}{A logical if is TRUE muestra graficos de densidad}
}
\value{
histogramas
}
\description{
Obtiene los histograma de cada columna
Convertir a data frame si se obtiene variables categoricas falsamente
}
\examples{
hist_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)),F)
}
