% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_col.R
\name{mean_col}
\alias{mean_col}
\title{Obtiene las medias por columnas
Convertir a data frame si se obtiene variables categoricas falsamente}
\usage{
mean_col(x)
}
\arguments{
\item{x}{A vector o dataset}
}
\value{
vector medias por columna
}
\description{
Obtiene las medias por columnas
Convertir a data frame si se obtiene variables categoricas falsamente
}
\examples{
mean_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)))
}
