% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNA.R
\name{replaceNA}
\alias{replaceNA}
\title{reemplaza los NAS
si al calcular y como media o mediana obiene error usar por ejemplo:
median(as.numeric(unlist(x)) ,na.rm = T)}
\usage{
replaceNA(x, y)
}
\arguments{
\item{x}{A vector o dataset}

\item{y}{el valor a reemplazar}
}
\value{
vector con NAS reemplazados
}
\description{
reemplaza los NAS
si al calcular y como media o mediana obiene error usar por ejemplo:
median(as.numeric(unlist(x)) ,na.rm = T)
}
\examples{
replaceNA(data.frame(x=c(1, 3, NA, NA, 6),y=c("A", NA, "C", "D", "E"),z=c(NA, TRUE, NA, TRUE, NA)),0)
}
