% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_col.R
\name{SD_col}
\alias{SD_col}
\title{Obtiene las desviaciones estandars por columnas
Convertir a data frame si se obtiene variables categoricas falsamente}
\usage{
SD_col(x)
}
\arguments{
\item{x}{A vector o dataset}
}
\value{
vector desviaciones estandars por columna
}
\description{
Obtiene las desviaciones estandars por columnas
Convertir a data frame si se obtiene variables categoricas falsamente
}
\examples{
SD_col(data.frame(x=c(1, 3, 2, 4, 6),y=c("A", "B", "C", "D", "E"),z=c(3, 4, 11, 15, 12)))
}
